[Update]Refactor Code
