--- src/Uiza.Net/.nuspec
@@ -1,46 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
-  <metadata>
-    <id>Uiza.net</id>
-    <version>1.0.0</version>
-    <authors>Uiza</authors>
-    <owners>Uiza</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>Uiza API Wrapper</description>
-    <dependencies>
-      <group targetFramework=".NETFramework4.5">
-        <dependency id="Newtonsoft.Json" version="12.0.1" exclude="Build,Analyzers" />
-        <dependency id="System.ComponentModel.Annotations" version="4.5.0" exclude="Build,Analyzers" />
-      </group>
-      <group targetFramework=".NETStandard1.6">
-        <dependency id="NETStandard.Library" version="1.6.1" exclude="Build,Analyzers" />
-        <dependency id="Newtonsoft.Json" version="12.0.1" exclude="Build,Analyzers" />
-        <dependency id="System.ComponentModel.Annotations" version="4.5.0" exclude="Build,Analyzers" />
-      </group>
-      <group targetFramework=".NETStandard2.0">
-        <dependency id="Newtonsoft.Json" version="12.0.1" exclude="Build,Analyzers" />
-        <dependency id="System.ComponentModel.Annotations" version="4.5.0" exclude="Build,Analyzers" />
-      </group>
-    </dependencies>
-    <frameworkAssemblies>
-      <frameworkAssembly assemblyName="Microsoft.CSharp" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Configuration" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Core" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Data.DataSetExtensions" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Data" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Data.Linq" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Net.Http" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Net.Http.WebRequest" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Web" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Xml" targetFramework=".NETFramework4.5" />
-      <frameworkAssembly assemblyName="System.Xml.Linq" targetFramework=".NETFramework4.5" />
-    </frameworkAssemblies>
-  </metadata>
-  
-   <files>
-    <file src="bin\Release\net45\**" target="lib\net45" />
-	  <file src="bin\Release\netstandard1.6\**" target="lib\netstandard1.6" />
-	  <file src="bin\Release\netstandard2.0\**" target="lib\netstandard2.0" />
-   </files>
-</package>
