+++ src/Uiza.Net/Configuration/UizaConfigOptions.cs
@@ -6,12 +6,13 @@
     public class UizaConfigOptions
     {
         /// <summary>
-        ///
+        /// Token get from API
+        /// https://docs.uiza.io/#get-api-key
         /// </summary>
         public string ApiKey { get; set; }
 
         /// <summary>
-        ///
+        /// Workspace API Domain
         /// </summary>
         public string ApiBase { get; set; }
     }
