+++ src/Uiza.Net/Configuration/UizaConfiguration.cs
@@ -23,13 +23,13 @@
         }
 
         /// <summary>
-        ///
+        /// Get Configuration Api Key
         /// </summary>
         /// <returns></returns>
         internal static string GetApiKey() => apiKey;
 
         /// <summary>
-        ///
+        /// Setup new API key
         /// </summary>
         /// <param name="newApiKey"></param>
         public static void SetApiKey(string newApiKey) => apiKey = newApiKey;
@@ -46,13 +46,13 @@
         }
 
         /// <summary>
-        ///
+        /// Setup new Workspace API Domain
         /// </summary>
         /// <param name="baseUrl"></param>
         public static void SetApiBase(string baseUrl) => apiBase = baseUrl;
 
         /// <summary>
-        ///
+        /// Setup Uiza Enviroment
         /// </summary>
         /// <param name="options"></param>
         public static void SetupUiza(UizaConfigOptions options)
