+++ src/Uiza.Net/Enum/CallbackEnums.cs
