+++ src/Uiza.Net/Enum/CategoryEnums.cs
