+++ src/Uiza.Net/Enum/StorageEnums.cs
