+++ src/Uiza.Net/Parameters/Analytic/AnalyticLineParameter.cs
@@ -8,13 +8,13 @@ namespace Uiza.Net.Parameters
     public class AnalyticLineParameter : BaseParameter
     {
         /// <summary>
-        ///
+        /// Start date (UTC+0) with format: YYYY-MM-DD
         /// </summary>
         [JsonProperty("start_date")]
         public string StartDate { get; set; }
 
         /// <summary>
-        ///
+        /// End date (UTC+0) with format: YYYY-MM-DD
         /// </summary>
         [JsonProperty("end_date")]
         public string EndDate { get; set; }
