+++ src/Uiza.Net/Parameters/Callback/UpdateCallbackParameter.cs
@@ -10,25 +10,25 @@ namespace Uiza.Net.Parameters
     public class UpdateCallbackParameter : RetrieveItemParameter
     {
         /// <summary>
-        ///
+        /// Your server URL for callback
         /// </summary>
         [JsonProperty("url")]
         public string Url { get; set; }
 
         /// <summary>
-        ///
+        /// Method of callback (get-post-put..)
         /// </summary>
         [JsonProperty("method")]
         public CallbackMethodTypes Method { get; set; }
 
         /// <summary>
-        ///
+        /// Extra data you want to attach in callback response	
         /// </summary>
         [JsonProperty("jsonData")]
         public Dictionary<string, object> JsonData { get; set; }
 
         /// <summary>
-        ///
+        /// Add parameter to headersAdd parameter to headers
         /// </summary>
         [JsonProperty("headersData")]
         public Dictionary<string, object> HeadersData { get; set; }
