+++ src/Uiza.Net/Parameters/Category/CategoryRelationParameter.cs
@@ -9,13 +9,13 @@ namespace Uiza.Net.Parameters
     public class CategoryRelationParameter : BaseParameter
     {
         /// <summary>
-        ///
+        /// Identifier of category
         /// </summary>
         [JsonProperty("metadataIds")]
         public List<string> MetadataIds { get; set; }
 
         /// <summary>
-        ///
+        /// Identifier of entity
         /// </summary>
         [JsonProperty("entityId")]
         public string EntityId { get; set; }
