+++ src/Uiza.Net/Parameters/LiveStreaming/UpdateLiveStreamingParameter.cs
@@ -24,6 +24,9 @@ namespace Uiza.Net.Parameters
 
         /// <summary>
         /// Type of event can be pull or push
+        /// Pull: We has supported RTMP, HLS and direct Live Youtube link. 
+        /// Uiza pull feed from pull link and broadcast it using Uiza's SDK. 
+        /// Push: Uiza give you a Publish endpoint, you can push feed into the endpoint and Uiza will broadcast it using Uiza's SDK.
         /// </summary>
         [JsonProperty("mode")]
         public string Mode { get; set; }
@@ -35,7 +38,7 @@ namespace Uiza.Net.Parameters
         public EncodeTypes Encode { get; set; }
 
         /// <summary>
-        ///
+        /// Resource mode ( single = only 1 feed and output), redundant = more than 1 feed and output to backup)
         /// </summary>
         [JsonProperty("resourceMode")]
         public ResourceModes ResourceMode { get; set; }
