--- src/Uiza.Net/Response/CustomResponse/CreateUserResponse.cs
@@ -1,16 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace Uiza.Net.Response
-{
-    /// <summary>
-    ///
-    /// </summary>
-    public class CreateUserResponse : UizaResponse
-    {
-        /// <summary>
-        ///
-        /// </summary>
-        [JsonProperty("id")]
-        public string Id { get; set; }
-    }
-}
